requirements.txt
loguru==0.7.0
streamlit==1.26.0
python-dotenv==1.0.0
Pillow~=9.5.0
requests~=2.31.0